component reorg
